+++ app/models/response_rate.rb
@@ -1,4 +1,7 @@
 module ResponseRate
+  TEACHER_RATE_THRESHOLD = 0.25
+  STUDENT_RATE_THRESHOLD = 0.25
+
   def initialize(subcategory:, school:, academic_year:)
     @subcategory = subcategory
     @school = school
@@ -13,5 +16,13 @@ module ResponseRate
     return 0 unless total_possible_responses.positive?
 
     (average_responses_per_survey_item / total_possible_responses.to_f * 100).round
+  end
+
+  def meets_student_threshold?
+    rate >= STUDENT_RATE_THRESHOLD * 100
+  end
+
+  def meets_teacher_threshold?
+    rate >= TEACHER_RATE_THRESHOLD * 100
   end
 end