+++ app/models/student_response_rate.rb
@@ -4,11 +4,11 @@ class StudentResponseRate
   private
 
   def survey_item_count
-    @student_survey_item_count ||= @subcategory.measures.map { |measure| measure.student_survey_items.count }.sum
+    @survey_item_count ||= @subcategory.measures.map { |measure| measure.student_survey_items.count }.sum
   end
 
   def response_count
-    @student_response_count ||= @subcategory.measures.map do |measure|
+    @response_count ||= @subcategory.measures.map do |measure|
       measure.student_survey_items.map do |survey_item|
         survey_item.survey_item_responses.where(school: @school, academic_year: @academic_year).count
       end.sum
@@ -16,7 +16,7 @@ class StudentResponseRate
   end
 
   def total_possible_responses
-    @total_possible_student_responses ||= begin
+    @total_possible_responses ||= begin
       total_responses = Respondent.where(school: @school, academic_year: @academic_year).first
       return 0 unless total_responses.present?