+++ app/models/subcategory.rb
@@ -10,4 +10,20 @@ class Subcategory < ActiveRecord::Base
     scores = scores.reject(&:nil?)
     scores.average
   end
+
+  def student_response_rate(school:, academic_year:)
+    @student_response_rate ||= Hash.new do |memo|
+      memo[[school, academic_year]] = StudentResponseRate.new(subcategory: self, school:, academic_year:)
+    end
+
+    @student_response_rate[[school, academic_year]]
+  end
+
+  def teacher_response_rate(school:, academic_year:)
+    @teacher_response_rate ||= Hash.new do |memo|
+      memo[[school, academic_year]] = TeacherResponseRate.new(subcategory: self, school:, academic_year:)
+    end
+
+    @teacher_response_rate[[school, academic_year]]
+  end
 end