+++ app/models/survey_item_response.rb
@@ -1,6 +1,6 @@
 class SurveyItemResponse < ActiveRecord::Base
-  TEACHER_RESPONSE_THRESHOLD = 17
-  STUDENT_RESPONSE_THRESHOLD = 196
+  TEACHER_RESPONSE_THRESHOLD = 2
+  STUDENT_RESPONSE_THRESHOLD = 2
 
   belongs_to :academic_year
   belongs_to :school