+++ app/models/teacher_response_rate.rb
@@ -12,11 +12,11 @@ class TeacherResponseRate
   end
 
   def survey_item_count
-    @teacher_survey_item_count ||= @subcategory.measures.map { |measure| measure.teacher_survey_items.count }.sum
+    @survey_item_count ||= @subcategory.measures.map { |measure| measure.teacher_survey_items.count }.sum
   end
 
   def response_count
-    @teacher_response_count ||= @subcategory.measures.map do |measure|
+    @response_count ||= @subcategory.measures.map do |measure|
       measure.teacher_survey_items.map do |survey_item|
         survey_item.survey_item_responses.where(school: @school, academic_year: @academic_year).count
       end.sum
@@ -24,7 +24,7 @@ class TeacherResponseRate
   end
 
   def total_possible_responses
-    @total_possible_teacher_responses ||= begin
+    @total_possible_responses ||= begin
       total_responses = Respondent.where(school: @school, academic_year: @academic_year).first
       return 0 unless total_responses.present?