+++ spec/system/journey_spec.rb
@@ -9,30 +9,36 @@ describe 'District Admin', js: true do
   let(:category) { Category.find_by_name('Teachers & Leadership') }
   let(:subcategory) { Subcategory.find_by_name('Teachers & The Teaching Environment') }
   let(:measures_for_subcategory) { Measure.where(subcategory:) }
-  let(:scales_for_subcategory) {Scale.where(measure: measures_for_subcategory)}
+  let(:scales_for_subcategory) { Scale.where(measure: measures_for_subcategory) }
   let(:survey_items_for_subcategory) { SurveyItem.where(scale: scales_for_subcategory) }
 
   let(:measure_1A_i) { Measure.find_by_measure_id('1A-i') }
   let(:measure_2A_i) { Measure.find_by_measure_id('2A-i') }
+  let(:measure_2A_ii) { Measure.find_by_measure_id('2A-ii') }
   let(:measure_4C_i) { Measure.find_by_measure_id('4C-i') }
   let(:measure_with_no_survey_responses) { Measure.find_by_measure_id('3A-i') }
 
-  let(:survey_item_1_for_measure_1A_i) { SurveyItem.create scale: measure_1A_i.scales.first, survey_item_id: rand.to_s }
-  let(:survey_item_2_for_measure_1A_i) { SurveyItem.create scale: measure_1A_i.scales.first, survey_item_id: rand.to_s }
-
   let(:survey_items_for_measure_1A_i) { measure_1A_i.survey_items }
   let(:survey_items_for_measure_2A_i) { measure_2A_i.survey_items }
+  let(:survey_items_for_measure_2A_ii) { measure_2A_ii.survey_items }
   let(:survey_items_for_measure_4C_i) { measure_4C_i.survey_items }
 
   let(:ay_2020_21) { AcademicYear.find_by_range '2020-21' }
-  let(:ay_2019_20) { AcademicYear.find_by_range '2019-20' }
+  # let(:ay_2019_20) { AcademicYear.find_by_range '2019-20' }
 
   let(:username) { 'winchester' }
   let(:password) { 'winchester!' }
+  let(:respondents) do
+    respondents = Respondent.where(school:).first
+    respondents.total_students = 8
+    respondents.total_teachers = 8
+    respondents.save
+  end
 
   before :each do
     Rails.application.load_seed
 
+    respondents
     survey_item_responses = []
 
     survey_items_for_measure_1A_i.each do |survey_item|
@@ -49,6 +55,13 @@ describe 'District Admin', js: true do
       end
     end
 
+    survey_items_for_measure_2A_ii.each do |survey_item|
+      SurveyItemResponse::STUDENT_RESPONSE_THRESHOLD.times do
+        survey_item_responses << SurveyItemResponse.new(response_id: rand.to_s, academic_year: ay_2020_21,
+                                                        school:, survey_item:, likert_score: 5)
+      end
+    end
+
     survey_items_for_measure_4C_i.each do |survey_item|
       SurveyItemResponse::TEACHER_RESPONSE_THRESHOLD.times do
         survey_item_responses << SurveyItemResponse.new(response_id: rand.to_s, academic_year: ay_2020_21,
@@ -57,7 +70,7 @@ describe 'District Admin', js: true do
     end
 
     survey_items_for_subcategory.each do |survey_item|
-      200.times do
+      2.times do
         survey_item_responses << SurveyItemResponse.new(response_id: rand.to_s, academic_year: ay_2020_21,
                                                         school:, survey_item:, likert_score: 4)
       end
@@ -108,6 +121,7 @@ end
 
 def district_admin_sees_student_physical_safety
   expect(page).to have_text('Student Physical Safety')
+
   expect(page).to have_css("[data-for-measure-id='2A-i'][width='40.0%'][x='60%']")
 end
 
@@ -159,7 +173,7 @@ def district_admin_sees_overview_content
   district_admin_sees_student_physical_safety
   district_admin_sees_problem_solving_emphasis
 
-  page.assert_selector('.measure-row-bar', count: 5)
+  page.assert_selector('.measure-row-bar', count: 6)
 end
 
 def district_admin_sees_browse_content